Added existing code used for the resin rocket testbench from the GUI repository
The diagram of the Ground Station and Electrical Junction Box shows the rough position of the internal circuit positions, external connectors so that we know where to drill holes on the physical boxes.

__init__.py tells Pycharm that packet.py is located in the current directory.

packet.py contains the packet the Arduino will use, converted to Python using ctypes.

main.py reads the USB port of the Arduino, displays a real time graph of the incoming data, and saves the data to an .csv file. It imports the Arduino's packet structure from packet.py.