Stored all known code into Ground Station GUI, Groundy, and Rocky folders.
Ground Station GUI contains the Python code that will run on a laptop connected to the ground station Arduino (Groundy) via USB.

Groundy contains all of the Arduino code for making the ground station work with the laptop, radio, and buttons/switches.

Rocky contains all of the Arduino code for controlling the rocket, reading sensors, and working with the radio.